Un-freeze Instrument dataclass, migrate the rest of the hierarchy
It's hard to make this frozen (at least, without a lot of hacks) because there are so many pre-conditions and post-conditions in these classes. Simpler to just let the fields be mutable, and trust users to not change them (or if they do, to understand how that will violate hashing).